+++ include/libkann/optimizers/SimpleOptimizer.hpp
@@ -10,12 +10,12 @@ namespace kann
   {
   public:
     KANN_EXPORT SimpleOptimizer() = default;
-    KANN_EXPORT SimpleOptimizer(double learningRate);
+    KANN_EXPORT SimpleOptimizer(float learningRate);
 
   public:
     KANN_EXPORT size_t process(Graph& graph, Info& info, Shape shape, size_t index, size_t gradient_index) const override;
 
   private:
-    double m_learningRate;
+    float m_learningRate;
   };
 }
